Moves /downloadTable to /responses
