+++ components/form-builder/app/responses/ConfirmReceiptStatus.tsx
